+++ optimizer/__init__.py
@@ -1,2 +1,1 @@
-from optimizer import *
-from tools import *
+""" Module with the optimizer and other suplementary functions"""
