--- optimizer/optimizer/__init__.py
@@ -1,2 +1,0 @@
-from minimizer import *
-from pyciminimizer import *
