--- optimizer/optimizer/tests/flow
@@ -1,4 +1,0 @@
-he_avdz_1.00.psi4.dat
-fci
-none
-end
