--- optimizer/optimizer/tests/run-pyci.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-PYTHONPATH=~/code/new-horton/:~/code/olsens/:~/code/olsens-new/:~/code/pyci1/:~/code/pyci/:~/mcmaster/EnergyExtrapolation/variational_pars/lroptimizer HORTONDATA=~/code/new-horton/data/ python
