--- optimizer/tools/__init__.py
@@ -1,1 +1,0 @@
-"""Module with tools and helpful functions."""
