+++ tools/__init__.py
@@ -1,0 +1,1 @@
+"""Module with tools and helpful functions."""
